update(gnome-extensions) refactor the whole source to separate components, hooks etc.
