+++ extensions/gnome-extensions/src/components/extension-list-item.tsx
@@ -1,0 +1,60 @@
+import React from "react";
+import { Action, ActionPanel, Clipboard, Icon, List, showToast, Toast } from "@vicinae/api";
+import useExtensionScreenshot from "../hooks/use-extension-screenshot";
+import { ExtensionListItemProps } from "../interfaces/extension-list-item-props";
+import { executeCommand } from "../utils/execute-command";
+import { ActionEnableExtension } from "./action-enable-extension";
+import { ActionDisableExtension } from "./action-disable-extension";
+import { ExtensionListDetail } from "./extension-list-detail";
+
+async function openExtensionPrefs(uuid: string): Promise<boolean> {
+  const { error } = await executeCommand(`gnome-extensions prefs "${uuid}"`);
+  return !error;
+}
+
+export function ExtensionListItem({ extension, isShowingDetail, onToggleDetail, onReload }: ExtensionListItemProps) {
+  const { screenshot, isLoading, openScreenshot } = useExtensionScreenshot(extension.uuid);
+
+  return (
+    <List.Item
+      key={extension.uuid}
+      title={extension.name}
+      subtitle={extension.description || extension.uuid}
+      icon={extension.enabled ? Icon.Checkmark : Icon.XMarkCircle}
+      accessories={extension.version ? [{ text: `v${extension.version}` }] : []}
+      detail={<ExtensionListDetail extension={extension} screenshot={screenshot} isLoadingScreenshot={isLoading} />}
+      actions={(
+        <ActionPanel>
+          <Action
+            title={isShowingDetail ? "Hide Details" : "Show Details"}
+            icon={Icon.Info01}
+            onAction={onToggleDetail}
+          />
+          <Action title="Preferences" icon={Icon.Cog} onAction={() => openExtensionPrefs(extension.uuid)} />
+          {screenshot && (
+            <Action
+              title="Open Screenshot"
+              icon={Icon.Image}
+              onAction={openScreenshot}
+              shortcut={{ modifiers: ["ctrl"], key: "i" }}
+            />
+          )}
+          {extension.enabled ? (
+            <ActionDisableExtension extension={extension} onReload={onReload} />
+          ) : (
+            <ActionEnableExtension extension={extension} onReload={onReload} />
+          )}
+          <Action
+            title="Copy UUID"
+            icon={Icon.CopyClipboard}
+            shortcut={{ modifiers: ["ctrl"], key: "u" }}
+            onAction={async () => {
+              await Clipboard.copy(extension.uuid);
+              await showToast({ style: Toast.Style.Success, title: "UUID copied" });
+            }}
+          />
+        </ActionPanel>
+      )}
+    />
+  );
+}
