+++ extensions/gnome-extensions/src/interfaces/extension-action.tsx
@@ -1,0 +1,6 @@
+import { GnomeExtension } from "./gnome-extension";
+
+export interface ExtensionAction {
+  extension: GnomeExtension;
+  onReload: () => void;
+}
