+++ extensions/gnome-extensions/src/interfaces/extension-detail-props.tsx
@@ -1,0 +1,7 @@
+import { GnomeExtension } from "./gnome-extension";
+
+export interface ExtensionDetailProps {
+  extension: GnomeExtension;
+  screenshot?: string;
+  isLoadingScreenshot: boolean;
+}
