+++ extensions/gnome-extensions/src/interfaces/extension-list-item-props.tsx
@@ -1,0 +1,8 @@
+import { GnomeExtension } from "./gnome-extension";
+
+export interface ExtensionListItemProps {
+  extension: GnomeExtension;
+  isShowingDetail: boolean;
+  onToggleDetail: () => void;
+  onReload: () => void;
+}
