+++ extensions/gnome-extensions/src/interfaces/extension-screenshot-props.tsx
@@ -1,0 +1,6 @@
+export interface ExtensionScreenshotProps {
+  screenshot?: string;
+  localPath?: string;
+  isLoading: boolean;
+  openScreenshot: () => void;
+}
