+++ extensions/gnome-extensions/src/interfaces/gnome-extension.tsx
@@ -1,0 +1,12 @@
+export interface GnomeExtension {
+  uuid: string;
+  name: string;
+  description: string;
+  enabled: boolean;
+  version?: string;
+  author?: string;
+  path?: string;
+  url?: string;
+  state?: string;
+  settingsSchema?: string;
+}
