+++ extensions/gnome-extensions/src/interfaces/preferences.tsx
@@ -1,0 +1,4 @@
+export interface Preferences {
+  extensionManagerPath?: string;
+  showDisabled?: boolean;
+}
