+++ extensions/gnome-extensions/src/utils/get-name-from-uuid.tsx
@@ -1,0 +1,4 @@
+export function getNameFromUuid(uuid: string): string {
+  const match = uuid.match(/^([a-zA-Z0-9_-]+)@/);
+  return match ? match[1] : uuid.split("@")[0];
+}
